feat:optimize menu management logic

+++ spring2go-common/spring2go-common-security/src/main/java/com/spring2go/common/security/domain/SimpleAuthorizeUser.java
@@ -14,13 +14,13 @@ import java.util.Set;
 public class SimpleAuthorizeUser implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private Integer userId;
+    private Long userId;
     private String username;
     private Object originalUser;
     private Set<String> roles;
     private Set<String> permissions;
 
-    public SimpleAuthorizeUser(Integer userId, String username, Object originalUser, Set<String> roles, Set<String> permissions) {
+    public SimpleAuthorizeUser(Long userId, String username, Object originalUser, Set<String> roles, Set<String> permissions) {
         this.userId = userId;
         this.username = username;
         this.originalUser = originalUser;
