+++ spring2go-modules/spring2go-system-biz/src/main/java/com/spring2go/system/service/impl/SysRoleMenuServiceImpl.java
@@ -30,7 +30,7 @@ public class SysRoleMenuServiceImpl extends ServiceImpl<SysRoleMenuMapper, SysRo
      */
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public Boolean saveRoleMenus(Integer roleId, String menuIds) {
+    public Boolean saveRoleMenus(Long roleId, String menuIds) {
 
         this.remove(Wrappers.<SysRoleMenu>query().lambda().eq(SysRoleMenu::getRoleId, roleId));
